Add files via upload
I Have used urllib library together with request class for url connection. Other than that in this code, I make use of BeautifulSoup library for extraction part . With the help of the library I got the html format data so it will be easily understood by user to how to get different output. I creted different methods for different question. I am here using procedural method of programmimg but we can use OOP method also. The The find_all method is used for finding out all tags with the specified tag name or id and returning them as a list of type bs4